requirements.txt
requests==2.28.1
jupyterlab==3.4.7
pyspark==3.2.1
PyYAML==6.0
ipykernel==6.15.0
boto3==1.24.82
pandas==1.3.5
pymongo[srv]
apache-airflow==2.4.3
python-dotenv
notebook
joblib==1.3.2
-e .
